MallBackend/requirements.txt
# 依赖列表
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-jwt-extended==4.5.3
python-dotenv==1.0.0
pymysql==1.1.0
requests==2.31.0
flask-swagger-ui==3.36.0
openai==1.30.1
flask-cors==4.0.0
cryptography==42.0.7